+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -5,7 +5,7 @@ object PluginVersions {
     val androidPlugin = "3.5.0"
     val semver = "1.1.1"
     val jgit = "5.6.0.201912101111-r"
-    val gradleWrapper = "6.0.1"
+    val gradleWrapper = "6.2.1"
     val junit5 = "5.5.2"
     val assertJ = "3.11.1"
     val shadow = "5.2.0"