+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -2,7 +2,7 @@ object PluginVersions {
     val kotlin = "1.3.60"
     val ktlintPlugin = "9.1.1"
     val gradlePublishPlugin = "0.10.1"
-    val androidPlugin = "3.5.0"
+    val androidPlugin = "3.6.0"
     val semver = "1.1.1"
     val jgit = "5.6.0.201912101111-r"
     val gradleWrapper = "6.2.1"