+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+### Added
+  - ?
+### Changed
+  - ?
+### Fixed
+  - ?
+### Removed
+  - ?
+
 ## [9.2.0] - 2020-02-10
 ### Added
   - Html reporter to provided reporters [#312](https://github.com/JLLeitschuh/ktlint-gradle/issues/312)
@@ -17,8 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - Shadow plugin dependencies into plugin jar
 ### Fixed
   - Fix `ktlintApplyToIdea` task fails when `android = true` is set [#311](https://github.com/JLLeitschuh/ktlint-gradle/issues/311)
-### Removed
-  - ?
 
 ## [9.1.1] - 2019-11-12
 ### Fixed